alloy_rlp

Trait MaxEncodedLen

pub unsafe trait MaxEncodedLen<const LEN: usize>: Encodable { }

Defines the max length of an Encodable type as a const generic.

Safety

An invalid value can cause the encoder to panic.