Trait alloy_rlp::Encodable

pub trait Encodable {
    // Required method
    fn encode(&self, out: &mut dyn BufMut);

    // Provided method
    fn length(&self) -> usize { ... }
}

A type that can be encoded via RLP.

Required Methods

fn encode(&self, out: &mut dyn BufMut)

Encodes the type into the out buffer.

Provided Methods

fn length(&self) -> usize

Returns the length of the encoding of this type in bytes.

The default implementation computes this by encoding the type. When possible, we recommender implementers override this with a specialized implementation.