alloy_rlp

Trait MaxEncodedLenAssoc

pub unsafe trait MaxEncodedLenAssoc: Encodable {
    const LEN: usize;
}

Defines the max length of an Encodable type as an associated constant.

Safety

An invalid value can cause the encoder to panic.

Required Associated Constants

const LEN: usize

The maximum length.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.